// infrastructure/src/main/webui/src/api/model/elementExecution.ts
/**
 * Generated by orval v6.24.0 🍺
 * Do not edit manually.
 * infrastructure API
 * OpenAPI spec version: 1.0-SNAPSHOT
 */
import type { Input } from "./input";
import type { Range } from "./range";

export interface ElementExecution {
  executedAtEnd: string;
  executedAtStart: string;
  inputs: Input[];
  processedInNanos: number;
  range: Range;
  result: string;
}